+++ R/joyn-package.R
@@ -23,7 +23,10 @@ if (getRversion() >= "2.15.1") {
       ":=",
       "..output",
       ".xreport",
-      ".yreport"
+      ".yreport",
+      'use_util_reportvar',
+      'varx_na',
+      'type'
     ),
     package = utils::packageName()
   )