+++ man/roxygen/meta.R
@@ -1,0 +1,4 @@
+list(
+  rd_family_title = list(options  = "JOYn options functions",
+                         messages = "Messages functions")
+)